move REALTEK_RTL8195AM.py, remove progen in targets.json
